requirements.txt
# Core dependencies
tensorflow==2.17.0
protobuf==3.20.3
numpy==1.26.0
streamlit
langchain
duckduckgo-search
langchain-community
langchain-groq
langgraph
openai
groq

# Voice interaction
pyttsx3
speechrecognition
pyaudio
pyautogui
psutil
winshell

# Web automation and utilities
pywhatkit
wikipedia

python-dotenv
